@@ -6,7 +6,7 @@ enable_testing()
 
 if (NOT ${TARGET} STREQUAL "")
   if (${TARGET} STREQUAL "server")
-      set(CMAKE_CXX_STANDARD 17)
+      set(CMAKE_CXX_STANDARD 20)
       set(CMAKE_CXX_STANDARD_REQUIRED ON)
       add_subdirectory(flatbuffers)
   endif()
